requirements.txt
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
coloredlogs==15.0.1
dotenv==0.9.9
flatbuffers==25.2.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.44.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==6.32.1
PyMatting==1.1.14
PyMuPDF==1.26.4
pytesseract==0.3.13
python-dotenv==1.1.1
python-telegram-bot==22.3
referencing==0.36.2
rembg==2.0.67
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.25.2
scipy==1.16.2
sniffio==1.3.1
sympy==1.14.0
tifffile==2025.9.9
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
